pyproject.toml
[project]
name = "new-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.10",
    "click>=8.2.1",
    "google-adk>=1.5.0",
    "google-genai>=1.24.0",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "uvicorn>=0.35.0",
]
